Dto ja luokat rokkaamaan

+++ Projektikansio/SharedLib/kuvaDTO.cs
@@ -2,9 +2,9 @@
 {
     public class kuvaDTO
     {
-        public int idkuva { get; set; }
+        public long idkuva { get; set; }
         public string kuva { get; set; }
-        public int idtarina { get; set; }
+        public long idtarina { get; set; }
 
     }
 }
